Make posterior predictive more robust. Make categorical a univariate family. Remove cylic import issues
