+++ bambi/families/family.py
@@ -1,8 +1,11 @@
 from typing import Dict, Union
 
+import numpy as np
+import pymc as pm
+import xarray as xr
+
 from bambi.families.link import Link
-from bambi.utils import get_auxiliary_parameters
-from bambi.posterior_predictive import get_posterior_predictive_draws
+from bambi.utils import get_auxiliary_parameters, get_aliased_name
 
 
 class Family:
@@ -103,16 +106,83 @@ class Family:
         self.default_priors.update(priors)
 
     def posterior_predictive(self, model, posterior, **kwargs):  # pylint: disable = unused-argument
-        """Get draws from posterior predictive distribution
+        """Get draws from the posterior predictive distribution
+
+        This function works for almost all the families. It grabs the draws for the parameters
+        needed in the response distribution, and then gets samples from the posterior predictive
+        distribution using `pm.draw()`. It won't work when the response distribution requires
+        parameters that are not available in `posterior`.
 
         Parameters
         ----------
         model : bambi.Model
             The model
-        posterior : xarray.Dataset
-            The dataset with draws for the parameters in the posterior distribution
+        posterior : xr.Dataset
+            The xarray dataset that contains the draws for all the parameters in the posterior.
+            It must contain the parameters that are needed in the distribution of the response, or
+            the parameters that allow to derive them.
+
+        Returns
+        -------
+        xr.DataArray
+            A data array with the draws from the posterior predictive distribution
         """
-        return get_posterior_predictive_draws(model, posterior)
+        response_dist = get_response_dist(model.family)
+        params = model.family.likelihood.params
+        response_aliased_name = get_aliased_name(model.response_component.response_term)
+
+        kwargs = {}
+        output_dataset_list = []
+
+        # In the posterior xr.Dataset we need to consider aliases.
+        # But we don't use aliases when passing kwargs to the PyMC distribution
+        for param in params:
+            # Extract posterior draws for the parent parameter
+            if param == model.family.likelihood.parent:
+                component = model.components[model.response_name]
+                var_name = f"{response_aliased_name}_mean"
+                kwargs[param] = posterior[var_name].to_numpy()
+                output_dataset_list.append(posterior[var_name])
+            else:
+                # Extract posterior draws for non-parent parameters
+                component = model.components[param]
+                component_aliased_name = component.alias if component.alias else param
+                var_name = f"{response_aliased_name}_{component_aliased_name}"
+                if var_name in posterior:
+                    kwargs[param] = posterior[var_name].to_numpy()
+                    output_dataset_list.append(posterior[var_name])
+                elif hasattr(component, "prior") and isinstance(component.prior, (int, float)):
+                    kwargs[param] = np.asarray(component.prior)
+
+        # Determine the array with largest number of dimensions
+        ndims_max = max(x.ndim for x in kwargs.values())
+
+        # Append a dimension when needed. Required to make `pm.draw()` work.
+        for key, values in kwargs.items():
+            kwargs[key] = expand_array(values, ndims_max)
+
+        # NOTE: Wouldn't it be better to always use parametrizations compatible with PyMC?
+        # The current approach allows more flexibility, but it's more painful.
+        if hasattr(model.family, "transform_backend_kwargs"):
+            kwargs = model.family.transform_backend_kwargs(kwargs)
+
+        output_array = pm.draw(response_dist.dist(**kwargs))
+        output_coords_all = xr.merge(output_dataset_list).coords
+
+        if hasattr(model.family, "KIND") and model.family.KIND == "Multivariate":
+            coord_names = (
+                "chain",
+                "draw",
+                response_aliased_name + "_obs",
+                response_aliased_name + "_mean_dim",
+            )
+        else:  # Assume it's univariate family
+            coord_names = ("chain", "draw", response_aliased_name + "_obs")
+
+        output_coords = {}
+        for coord_name in coord_names:
+            output_coords[coord_name] = output_coords_all[coord_name]
+        return xr.DataArray(output_array, coords=output_coords)
 
     def __str__(self):
         msg_list = [f"Family: {self.name}", f"Likelihood: {self.likelihood}", f"Link: {self.link}"]
@@ -120,3 +190,47 @@ class Family:
 
     def __repr__(self):
         return self.__str__()
+
+
+def get_response_dist(family):
+    """Get the PyMC distribution for the response
+
+    Parameters
+    ----------
+    family : bambi.Family
+        The family for which the response distribution is wanted
+
+    Returns
+    -------
+    pm.Distribution
+        The response distribution
+    """
+    if family.likelihood.dist:
+        dist = family.likelihood.dist
+    else:
+        dist = getattr(pm, family.likelihood.name)
+    return dist
+
+
+def expand_array(x, ndim):
+    """Add dimensions to an array to match the number of desired dimensions
+
+    If x.ndim < ndim, it adds ndim - x.ndim dimensions after the last axis. If not, it is left
+    untouched.
+
+    Parameters
+    ----------
+    x : np.ndarray
+        The array
+    ndim : int
+        The number of desired dimensions
+
+    Returns
+    -------
+    np.ndarray
+        The array with the expanded dimensions
+    """
+    if x.ndim == ndim:
+        return x
+    dims_to_expand = tuple(range(ndim - 1, x.ndim - 1, -1))
+    return np.expand_dims(x, dims_to_expand)
