+++ bambi/families/multivariate.py
@@ -1,55 +1,14 @@
 # pylint: disable=unused-argument
-import pytensor.tensor as pt
 import numpy as np
 import xarray as xr
+import pytensor.tensor as pt
 
 from bambi.families.family import Family
 from bambi.utils import extract_argument_names, extra_namespace, get_aliased_name
 
 
 class MultivariateFamily(Family):
-    pass
-
-
-class Categorical(MultivariateFamily):
-    SUPPORTED_LINKS = {"p": ["softmax"]}
-    UFUNC_KWARGS = {"axis": -1}
-
-    def transform_linear_predictor(self, model, linear_predictor):
-        response_name = get_aliased_name(model.response_component.response_term)
-        response_levels_dim = response_name + "_dim"
-        linear_predictor = linear_predictor.pad({response_levels_dim: (1, 0)}, constant_values=0)
-        return linear_predictor
-
-    def transform_coords(self, model, mean):
-        # The mean has the reference level in the dimension, a new name is needed
-        response_name = get_aliased_name(model.response_component.response_term)
-        response_levels_dim = response_name + "_dim"
-        response_levels_dim_complete = response_name + "_mean_dim"
-        levels_complete = model.response_component.response_term.levels
-        mean = mean.rename({response_levels_dim: response_levels_dim_complete})
-        mean = mean.assign_coords({response_levels_dim_complete: levels_complete})
-        return mean
-
-    def get_data(self, response):
-        return np.nonzero(response.term.data)[1]
-
-    def get_coords(self, response):
-        name = response.name + "_dim"
-        return {name: [level for level in response.levels if level != response.reference]}
-
-    def get_reference(self, response):
-        return get_reference_level(response.term)
-
-    @staticmethod
-    def transform_backend_nu(nu, data):
-        # Add column of zeros to the linear predictor for the reference level (the first one)
-        shape = (data.shape[0], 1)
-
-        # The first line makes sure the intercept-only models work
-        nu = np.ones(shape) * nu  # (response_levels, ) -> (n, response_levels)
-        nu = pt.concatenate([np.zeros(shape), nu], axis=1)
-        return nu
+    KIND = "Multivariate"
 
 
 class Multinomial(MultivariateFamily):
@@ -137,19 +96,3 @@ class Multinomial(MultivariateFamily):
         nu = np.ones(shape) * nu  # (response_levels, ) -> (n, response_levels)
         nu = pt.concatenate([np.zeros(shape), nu], axis=1)
         return nu
-
-
-# pylint: disable = protected-access
-def get_reference_level(term):
-    if term.kind != "categoric":
-        return None
-
-    if term.levels is None:
-        return None
-
-    levels = term.levels
-    intermediate_data = term.components[0]._intermediate_data
-    if hasattr(intermediate_data, "_contrast"):
-        return intermediate_data._contrast.reference
-
-    return levels[0]
